URL and Email external libraries + XML conversion
Added external libraries for URL and Email validation. This will allow for much more robust identification of URLs and Emails. Useful for URL quarantining and Sender validation respectively.

Also figured out the XML format for inputting data and created a function to convert the input string to XML and to navigate through the XML tags. This should simplify input validation, allow for seamless integration with a potential XML input file and make it easier to separate the message body into different sections for parsing.

I now posses all the required libraries and have a clear idea of how to complete the prototype application.